CHE-84 Post controller directory restructure
