+++ server/controllers/postController/index.ts
@@ -1,0 +1,6 @@
+import listPostsByThreadId from './listPostsByThreadId/listPostsByThreadId';
+import createPost from './createPost/createPost';
+import updatePost from './updatePost/updatePost';
+import deletePost from './deletePost/deletePost';
+
+export { listPostsByThreadId, createPost, updatePost, deletePost };
